chore: reorganize config folder structure to comply with standards
