--- src/config/creditLines/addCreditLine/navigation.ts
@@ -1,21 +1,0 @@
-const crumbsAddCreditLines = [
-  {
-    path: "/",
-    label: "Inicio",
-    id: "home",
-    isActive: false,
-  },
-  {
-    path: "/credit-lines",
-    label: "Líneas de crédito",
-    id: "creditLine",
-    isActive: true,
-  },
-  {
-    path: "/credit-lines/add-credit-line",
-    label: "Agregar Linea de Crédito",
-    id: "addCreditLine",
-    isActive: true,
-  },
-];
-export { crumbsAddCreditLines };
