--- src/config/creditLines/details/tabs.config.tsx
@@ -1,24 +1,0 @@
-const detailsTabsConfig = {
-  generalData: {
-    id: "generalData",
-    isDisabled: false,
-    label: "1. Datos generales",
-  },
-  creditProspectusOptions: {
-    id: "creditProspectusOptions",
-    isDisabled: false,
-    label: "2. Opciones para el prospecto de crédito",
-  },
-  maximumAmount: {
-    id: "maximumAmount",
-    isDisabled: false,
-    label: "3. Monto máximo",
-  },
-  term: {
-    id: "term",
-    isDisabled: false,
-    label: "4. Plazo",
-  },
-};
-
-export { detailsTabsConfig };
