+++ src/config/menuMainConfiguration/index.tsx
@@ -1,0 +1,19 @@
+import { MdLogout } from "react-icons/md";
+
+const userMenu = [
+    {
+      id: "section",
+      title: "",
+      links: [
+        {
+          id: "logout",
+          title: "Cerrar sesión",
+          path: "/logout",
+          iconBefore: <MdLogout />,
+        },
+      ],
+      divider: true,
+    },
+  ];
+  
+export { userMenu };
