+++ src/config/moneyDestination/moneyDestinationTab/generics/deleteMoneyDestinationModal/index.tsx
