+++ src/config/moneyDestination/moneyDestinationTab/generics/detailsTabsConfig/index.tsx
@@ -1,0 +1,15 @@
+const detailsTabsConfig = {
+  generalData: {
+    id: "generalData",
+    isDisabled: false,
+    label: "Datos generales",
+  },
+  creditLine: {
+    id: "creditLine",
+    isDisabled: false,
+    label: "Línea de crédito",
+    notificationIndicators: 2,
+  },
+};
+
+export { detailsTabsConfig };
