+++ src/config/moneyDestination/requestsInProgressTab/details/labelsOfRequest/index.ts
@@ -1,0 +1,16 @@
+const labelsOfRequest = [
+    {
+      id: "date",
+      titleName: "Fecha de solicitud:",
+    },
+    {
+      id: "responsible",
+      titleName: "Responsable:",
+    },
+    {
+      id: "status",
+      titleName: "Estado:",
+    },
+  ];
+
+export { labelsOfRequest };
