+++ src/config/moneyDestination/requestsInProgressTab/details/labelsOfTraceability/index.ts
@@ -1,0 +1,8 @@
+const labelsOfTraceability = [
+    {
+      id: "observation",
+      titleName: "Observación:",
+    },
+  ];
+
+export { labelsOfTraceability };
