--- src/config/moneyDestination/requestsInProgressTab/details/parameters.config.tsx
@@ -1,23 +1,0 @@
-const labelsOfRequest = [
-  {
-    id: "date",
-    titleName: "Fecha de solicitud:",
-  },
-  {
-    id: "responsible",
-    titleName: "Responsable:",
-  },
-  {
-    id: "status",
-    titleName: "Estado:",
-  },
-];
-
-const labelsOfTraceability = [
-  {
-    id: "observation",
-    titleName: "Observación:",
-  },
-];
-
-export { labelsOfRequest, labelsOfTraceability };
