+++ src/pages/creditLines/interface.tsx
@@ -12,7 +12,7 @@ import { ComponentAppearance } from "@ptypes/aparences.types";
 import { dataCreditLines } from "@mocks/creditLines/creditLines.mock";
 import { StyledContent } from "./styles";
 import { crumbsCreditLines } from "@config/creditLines/navigation";
-import { actions, breakPoints, titles } from "@config/creditLines/table.config";
+import { actions, breakPoints, titles } from "@config/creditLines/table";
 
 
 interface ICreditLinesUI {
