+++ src/pages/home/index.tsx
@@ -1,9 +1,9 @@
 import { useContext, useEffect, useRef, useState } from "react";
-import { appCards } from "@config/appCards";
 
 import { IBusinessUnitsPortalStaff } from "@ptypes/staffPortalBusiness.types";
 import { HomeUI } from "./interface";
 import { AuthAndPortalData } from "@context/authAndPortalDataProvider";
+import { mainCards } from "@config/mainCard";
 
 function Home() {
   const { appData, businessUnitsToTheStaff, setBusinessUnitSigla } =
@@ -33,7 +33,7 @@ function Home() {
       businessUnitsToTheStaff={businessUnitsToTheStaff}
       collapse={collapse}
       collapseMenuRef={collapseMenuRef}
-      data={appCards}
+      data={mainCards}
       selectedClient={selectedClient}
       setCollapse={setCollapse}
       handleLogoClick={handleLogoClick}
