+++ src/pages/home/interface.tsx
@@ -4,7 +4,7 @@ import { Icon } from "@inubekit/icon";
 import { useMediaQueries } from "@inubekit/hooks";
 
 import { AppCard } from "@components/feedback/AppCard";
-import { nav, userMenu } from "@config/nav";
+
 import { Title } from "@components/data/Title";
 import { BusinessUnitChange } from "@design/inputs/BusinessUnitChange";
 import { IBusinessUnitsPortalStaff } from "@ptypes/staffPortalBusiness.types";
@@ -23,6 +23,8 @@ import {
   StyledTitle,
 } from "./styles";
 import { IAppData } from "@context/authAndPortalDataProvider/types";
+import { mainNavigation } from "@config/mainNavigation";
+import { userMenu } from "@config/menuMainConfiguration";
 
 interface IHomeUI {
   appData: IAppData;
@@ -74,7 +76,7 @@ function HomeUI(props: IHomeUI) {
         <StyledHeaderContainer>
           <Header
             portalId="portal"
-            navigation={nav}
+            navigation={mainNavigation}
             logoURL={renderLogo(appData.businessUnit.urlLogo)}
             user={{
               username: appData.user.userName,
