+++ src/pages/moneyDestination/tabs/moneyDestinationTab/forms/creditLine/interface.tsx
@@ -13,7 +13,7 @@ import {
   deleteModal,
 } from "@design/forms/decisions/config/messages.config";
 import { StyledContainer } from "./styles";
-import { textValuesBusinessRules } from "@config/moneyDestination/moneyDestinationTab/businessRules.config";
+import { textValuesBusinessRules } from "@config/moneyDestination/moneyDestinationTab/businessRules";
 
 interface ICreditLineFormUI {
   decisions: IRuleDecision[];
