+++ src/pages/moneyDestination/tabs/moneyDestinationTab/interface.tsx
@@ -9,7 +9,7 @@ import { ComponentAppearance } from "@ptypes/aparences.types";
 import { Table } from "@components/data/Table";
 import { IEntry } from "@components/data/Table/types";
 import { StyledContainer } from "./styles";
-import { actions, breakPoints, titles } from "@config/moneyDestination/moneyDestinationTab/table.config";
+import { actions, breakPoints, titles } from "@config/moneyDestination/moneyDestinationTab/table";
 
 interface IMoneyDestinationTabUI {
   entries: IEntry[];
